KH Classify: extract source/type slide-in form into dialog component
ClassifySourceTypeFormDialogComponent owns all Form*/dropdown/dynamic-field
state + SaveSource/SaveContentType, reading metadata from
KnowledgeHubMetadataEngine/AIEngineBase directly. Host delegates via @ViewChild
(tab Add/Edit contracts unchanged), reloads on Saved, triggers no-content-type
warning via ContentTypeMissing output. Host 2157 -> 1616 TS lines.

DECOMPOSITION COMPLETE: host 5147 -> 1616 TS (69% smaller); 6 tab components +
3 dialog components; remaining dialogs co-located with their owning tabs. Build
green.
