+++ packages/Angular/Explorer/dashboards/src/AI/components/autotagging/dialogs/source-type-form.dialog.component.css
@@ -1,0 +1,8 @@
+/*
+ * Source + Content Type CRUD slide-in form dialog styles.
+ *
+ * The host (AutotaggingPipelineResourceComponent) uses ViewEncapsulation.None,
+ * so its `.at-slide-*` / `.at-form-*` rules are global and already style this
+ * dialog's markup while it renders inside the host shell. No dialog-specific
+ * rules are needed today.
+ */
